+++ ceno_emul/src/syscalls/bn254/mod.rs
@@ -1,0 +1,12 @@
+mod bn254;
+mod bn254_fptower;
+mod types;
+
+pub use bn254::{Bn254AddSpec, Bn254DoubleSpec, bn254_add, bn254_double};
+
+pub use bn254_fptower::{
+    Bn254Fp2AddSpec, Bn254Fp2MulSpec, Bn254FpAddSpec, Bn254FpMulSpec, bn254_fp_add, bn254_fp_mul,
+    bn254_fp2_add, bn254_fp2_mul,
+};
+
+pub use types::{BN254_FP_WORDS, BN254_FP2_WORDS, BN254_POINT_WORDS};
