+++ examples/Cargo.toml
@@ -15,5 +15,5 @@ rkyv = { version = "0.8", default-features = false, features = [
   "alloc",
   "bytecheck",
 ] }
-num-bigint.workspace = true
+# rand.workspace = true
 substrate-bn.workspace = true
